Equipment selection rewrite.
There is still a lot of work to be done with this and there are probably
various bugs, but it seems to have basic functionality at this point. The
idea of this rewrite is that qw should pick up and carry around items that
it thinks it might want in the future (when resists change or it enters a
different branch) even if it isn't going to use them at the moment. This
allows for things like swapping for rF for Zot or swapping for rCorr
for Slime (not yet implemented) and there are various other things I can
imagine implementing this way (swapping away from vamp on the orbrun,
waiting to use exec axes until axes skill is high enough, etc).

The rewrite also makes a start at handling artefact and ego properties
better, though there are many properties that still aren't tracked.